feat: navigate함수 사용
