+++ src/components/ui/modal/Modal.css
@@ -4,11 +4,12 @@
   left: 0;
   width: 100%;
   height: 100%;
-  background-color: rgba(0, 0, 0, 0.5); /* 반투명 배경 */
-  display: none; /* 기본적으로 숨김 */
+  background-color: rgba(0, 0, 0, 0.5);
+  display: none;
   justify-content: center;
   align-items: center;
-  z-index: 1000; /* 다른 요소들 위에 표시 */
+  z-index: 1000;
+  transition: opacity 0.3s ease;
 }
 
 .modal-content {
@@ -17,14 +18,14 @@
   border-radius: 5px;
   position: relative;
   width: 500px;
-  height: 300px; /* 높이를 자동으로 조정하여 내용에 맞춤 */
-  margin: auto; /* 중앙 정렬을 위한 마진 */
-  box-shadow: 0 4px 8px rgba(0, 0, 0, 0.2); /* 그림자 효과 추가 */
-  display: flex; /* Flexbox로 설정 */
-  flex-direction: column; /* 세로 방향으로 정렬 */
-  align-items: center; /* 중앙 정렬 */
+  height: 320px;
+  margin: auto;
+  box-shadow: 0 4px 8px rgba(0, 0, 0, 0.2);
+  display: flex;
+  flex-direction: column;
+  align-items: center;
 }
 
 .modal.show {
-  display: flex; /* flexbox로 설정하여 중앙 정렬 */
+  display: flex;
 }
