+++ src/components/ui/modal/user/userModal.css
@@ -4,81 +4,82 @@ button[data-type='punch-in'],
 button[data-type='punch-out'] {
   width: 100px;
   height: 50px;
-  background-color: var(--color-skyblue-light); /* 출근 버튼의 배경색 */
+  background-color: var(--color-skyblue-light);
   margin-bottom: 30px;
 }
 
 .cancel-button {
   background-color: var(--color-coral);
   width: 100px;
-  height: 50px; /* 취소 버튼의 배경색 */
+  height: 50px;
 }
 
 .modal-title {
   font-size: 32px;
   font-weight: 700;
-  margin: 35px 0; /* 제목과 내용 사이의 간격 */
+  margin: 35px 0;
 }
 
 .modal-date,
 .modal-time {
   font-size: 24px;
-  margin: 10px 20px; /* 날짜와 시간 사이의 간격 */
+  margin: 10px 20px;
 }
 .modal-time {
   margin-left: 30px;
 }
 
 .button-container {
-  display: flex; /* 버튼을 가로로 나열 */
-  justify-content: space-between; /* 버튼 사이에 공간을 둡니다 */
-  margin-top: 20px; /* 버튼과 이전 요소 사이의 간격 */
-  width: 100%; /* 버튼 컨테이너의 너비를 100%로 설정 */
+  display: flex;
+  justify-content: space-between;
+  margin-top: 20px;
+  width: 100%;
 }
 
 /* modal의 버튼에만 해당 */
 .modal button {
-  padding: 10px 20px; /* 버튼 패딩 */
-  font-size: 16px; /* 버튼 폰트 크기 */
-  border: none; /* 기본 테두리 제거 */
-  border-radius: 5px; /* 모서리 둥글게 */
-  cursor: pointer; /* 커서 모양 변경 */
-  margin-bottom: 20px; /* 버튼 아래 간격 추가 */
+  padding: 10px 20px;
+  font-size: 16px;
+  border: none;
+  border-radius: 5px;
+  cursor: pointer;
+  margin-bottom: 20px;
 }
 
 /* 성공한 경우 */
 
 /* 성공 메시지 스타일 */
 .success-message {
-  text-align: center; /* 중앙 정렬 */
-  font-size: 1.2em; /* 글자 크기 */
-  margin-top: 20px; /* 위 여백 */
+  text-align: center;
+  font-size: 1.2em;
+  margin-top: 20px;
 }
 
 .success-message p {
-  line-height: 1.5; /* 줄 간격을 1.5로 설정 */
+  line-height: 1.5;
 }
 
 .modal-content .material-symbols-rounded {
-  font-size: 150px; /* 원하는 크기로 수정 */
-  display: block; /* 블록으로 표시 */
-  margin: 0 auto; /* 중앙 정렬 */
+  font-size: 150px;
+  display: block;
+  margin: 0 auto;
 }
 .success-message .close-button {
   width: 100px;
   height: 40px;
+  margin-top: 10px;
   background-color: var(--color-light-gray);
 }
 
 /* 실패한 경우 */
 .error-message {
-  text-align: center; /* 중앙 정렬 */
-  font-size: 1.2em; /* 글자 크기 */
-  margin-top: 20px; /* 위 여백 */
+  text-align: center;
+  font-size: 1.2em;
+  margin-top: 20px;
 }
 
 .error-message p {
-  line-height: 1.5; /* 줄 간격을 1.5로 설정 */
+  line-height: 1.5;
 }
 
 .error-message .retry-button,
